requirements.txt
scipy~=1.14.1
opencv-python~=4.10.0.84
numpy~=1.26.4
pandas~=2.2.3
matplotlib~=3.9.2
scikit-learn~=1.5.2
torchvision~=0.19.1
setuptools~=75.1.0
yaml~=0.2.5
pyyaml~=6.0.2